notify home page check to mail

+++ Pro.Service/Pro.Service/Implements/GetDataService.cs
@@ -42,43 +42,50 @@ namespace Pro.Service.Implements
                     statusGetData2 = false;
                     LogHelper.Info($"GET---Start GetDataService");
                     string localPath = "";
-                    var newestChapDatas = _prepareService.PrepareNewestChapDatasForNew(ref localPath);
-                    if (newestChapDatas.Chaps != null && newestChapDatas.Chaps.Any())
+                    if (_prepareService.IsValidHomePage(true))
                     {
-                        if (_getRawDataService.GetRawDatasForNew(newestChapDatas))
+                        var newestChapDatas = _prepareService.PrepareNewestChapDatasForNew(ref localPath);
+                        if (newestChapDatas.Chaps != null && newestChapDatas.Chaps.Any())
                         {
-                            //SaveData2File($@"D:\Debug\RawData{newestChapDatas.Name}.json", newestChapDatas);
-
-                            //Save to file
-                            //newestChapDatas = ReadDataFromFile($@"D:\Debug\RawData{newestChapDatas.Name}.json");
-                            //LogHelper.Info($"GET---Start SaveImage2LocalFunc");
-                            //_saveImage2Local.SaveImage2LocalFunc(newestChapDatas);
-                            //SaveData2File($@"D:\Debug\SavedLocal_{newestChapDatas.Name}.json", newestChapDatas);
-
-                            //Uplpad to Clound
-                            //newestChapDatas = ReadDataFromFile($@"D:\Debug\SavedLocal_{newestChapDatas.Name}.json");
-                            if (_uploadImageService.HasValidCloudinary(isNotify: true))
+                            if (_getRawDataService.GetRawDatasForNew(newestChapDatas))
                             {
-                                LogHelper.Info($"GET---Start UploadLink2StoreWith3ThreadsForNew");
-                                _uploadImageService.UploadLink2StoreWith3ThreadsForNew(newestChapDatas);
-                                //SaveData2File($@"D:\Debug\HasClound_{newestChapDatas.Name}.json", newestChapDatas);
-                                //Save to DB
-                                //_upData2DBService.UpData2DBForNew(newestChapDatas);
-                                //Delete file
-                                try
+                                //SaveData2File($@"D:\Debug\RawData{newestChapDatas.Name}.json", newestChapDatas);
+
+                                //Save to file
+                                //newestChapDatas = ReadDataFromFile($@"D:\Debug\RawData{newestChapDatas.Name}.json");
+                                //LogHelper.Info($"GET---Start SaveImage2LocalFunc");
+                                //_saveImage2Local.SaveImage2LocalFunc(newestChapDatas);
+                                //SaveData2File($@"D:\Debug\SavedLocal_{newestChapDatas.Name}.json", newestChapDatas);
+
+                                //Uplpad to Clound
+                                //newestChapDatas = ReadDataFromFile($@"D:\Debug\SavedLocal_{newestChapDatas.Name}.json");
+                                if (_uploadImageService.HasValidCloudinary(isNotify: true))
                                 {
-                                    FileReader.DeleteFile(localPath);
+                                    LogHelper.Info($"GET---Start UploadLink2StoreWith3ThreadsForNew");
+                                    _uploadImageService.UploadLink2StoreWith3ThreadsForNew(newestChapDatas);
+                                    //SaveData2File($@"D:\Debug\HasClound_{newestChapDatas.Name}.json", newestChapDatas);
+                                    //Save to DB
+                                    _upData2DBService.UpData2DBForNew(newestChapDatas);
+                                    //Delete file
+                                    try
+                                    {
+                                        FileReader.DeleteFile(localPath);
+                                    }
+                                    catch (Exception ex)
+                                    {
+                                        LogHelper.Error($"DeleteFile/Move {localPath}" + ex);
+                                    }
                                 }
-                                catch (Exception ex)
+                                else
                                 {
-                                    LogHelper.Error($"DeleteFile/Move {localPath}" + ex);
+                                    LogHelper.Error($"No valid cloudinary!");
                                 }
-                            }
-                            else
-                            {
-                                LogHelper.Error($"No valid cloudinary!");
                             }
                         }
+                    }
+                    else
+                    {
+                        LogHelper.Warn($"Invalid data HomePage url");
                     }
                     statusGetData2 = true;
                     LogHelper.Info($"GET---Stop GetDataService");
